pyproject.toml
[project]
name = "training-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "lightning>=2.4",
    "torch>=2.2",
    "pyyaml>=6.0",
    "tqdm>=4.0",
    "python-decouple>=3.8",
    "wonderwords>=3.0",
    "rich>=13.0",
]

[project.optional-dependencies]
# Use in Colab:
# %pip install "training-template[colab] @ git+https://github.com/quentin-auster/dl-training-template.git"
colab = [
    "numpy>=2.0",
    "matplotlib>=3.8",
    "torchinfo>=1.8",
]

train = [
    "wandb>=0.18",
    "tensorboard>=2.15",
]

asr = [
    "torchaudio>=2.2",
    "jiwer>=3.0",
]

[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

analysis = [
    "jupyterlab>=4.5.3",
]
